+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_response_handlers.py
@@ -36,6 +36,7 @@ from ._models import (
     RecognizeCustomEntitiesResult,
     ClassifyDocumentResult,
     ActionPointerKind,
+    ExtractSummaryResult,
 )
 
 
@@ -262,6 +263,15 @@ def healthcare_result(
 
 
 @prepare_result
+def summary_result(
+    summary, results, *args, **kwargs
+):  # pylint: disable=unused-argument
+    return ExtractSummaryResult._from_generated(  # pylint: disable=protected-access
+        summary
+    )
+
+
+@prepare_result
 def custom_entities_result(
     custom_entities, results, *args, **kwargs
 ):  # pylint: disable=unused-argument
@@ -312,6 +322,8 @@ def _get_deserialization_callback_from_task_type(task_type):  # pylint: disable=
         return classify_document_result
     if task_type == _AnalyzeActionsType.ANALYZE_HEALTHCARE_ENTITIES:
         return healthcare_result
+    if task_type == _AnalyzeActionsType.EXTRACT_SUMMARY:
+        return summary_result
     return key_phrases_result